fix(dispatcher): use Double instead of Float for time precision
Float (~7 significant digits) loses precision at current Unix epoch
(~1.7e9 seconds), causing up to ~200 second timing errors.

Double (~15 significant digits) provides adequate precision for
millisecond-level time calculations.

Addresses PR #255 review feedback issue #24.

+++ core/service/Service/Integration/Dispatcher.hs
@@ -69,6 +69,7 @@ import ConcurrentVar (ConcurrentVar)
 import ConcurrentVar qualified
 import Console qualified
 import Data.Time.Clock.POSIX qualified as GhcPosix
+import GHC.Float (Double)
 import GHC.Real qualified as GhcReal
 import Integration qualified
 import Json qualified
@@ -360,7 +361,7 @@ getOrCreateLifecycleWorker dispatcher streamId = do
 getCurrentTimeMs :: Task Text Int
 getCurrentTimeMs = do
   posixTime <- GhcPosix.getPOSIXTime |> Task.fromIO
-  let millis = round (GhcReal.realToFrac posixTime * 1000 :: Float)
+  let millis = round (GhcReal.realToFrac posixTime * 1000 :: Double)
   Task.yield millis
